Renamed the accreditations module to memberships
