+++ docs/.vuepress/config.js
@@ -48,7 +48,6 @@ module.exports = {
                 title: "Modules",
                 collapsable: false,
                 children: [
-                    "x/accreditations/",
                     "x/memberships/",
                     "x/docs/",
                     "x/government/",
