--- docs/x/accreditations/README.md
@@ -1,23 +1,0 @@
-# Accreditations 
-Inside Commercio.network we've designed a system that allows you to invite users and later get rewarded when the 
-buy a new membership.
-
-## Accreditation procedure
-The procedure that allows a user to be accreditated is composed of two processes:
-
-- [Invitation process](invitation-process.md)
-- [Verification process](verification-process.md)
-
-These processes can be carried out independently one from the other, and in any order you desire.  
-
-## Transactions
-Using the `accreditations` module you can perform the following transactions. 
-
-**Accessible to everyone**
-- [Inviting a user](tx/inviting-user.md)
-
-**Accessible to TSPs**
-- [Marking a user as verified](tx/verify-user.md)
-
-**Accessible only to the government**
-- [Recognizing a TSP](tx/recognizing-a-tsp.md)
