+++ docs/x/memberships/invitation-process.md
@@ -1,5 +1,5 @@
 # Invitation process
-The invitation process is the first su-process of the [accreditation procedure](README.md#accreditation-procedure) 
+The invitation process is the first su-process of the accreditation procedure. 
 which allows you to invite a new user to join Commercio.network so that you can later be rewarded if he later
 buys a membership.  
 
@@ -7,7 +7,7 @@ buys a membership.
 Suppose that you want to invite another user and later receive a reward once he buys a membership. 
 Then, the following requirements must be met:
 
-* You must already have [bought](../memberships/tx/buy-membership.md) a membership. 
+* You must already have [bought](tx/buy-membership.md) a membership. 
 * You must know the user's identity that he has created by following the [specific guide](../id/tx/create-an-identity.md).
 
 ## Inviting a user
