+++ docs/x/memberships/tx/recognizing-a-tsp.md
