+++ docs/x/memberships/verification-process.md
@@ -1,8 +1,7 @@
 # Verification process
 The verification process allows any Trusted Service Provider (*TSP*) to mark a user as being already verified.  
 This procedure needs to be followed in order to allow any user to buy a membership and is the second sub-process 
-(along with the [invitation one](invitation-process.md)) to be part of the
-[accreditation procedure](README.md#accreditation-procedure).
+(along with the [invitation one](invitation-process.md)) to be part of the accreditation procedure.
 
 ## Verifying a user
 :::warning
